+++ Philosophers/utils.c
@@ -6,17 +6,17 @@
 /*   By: lelanglo <lelanglo@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/28 09:46:46 by lelanglo          #+#    #+#             */
-/*   Updated: 2025/01/30 15:15:40 by lelanglo         ###   ########.fr       */
+/*   Updated: 2025/02/01 17:12:05 by lelanglo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	i;
-	int	result;
-	int	sign;
+	int		i;
+	long	result;
+	int		sign;
 
 	i = 0;
 	result = 0;
